feat(hw3): add role requirements
