+++ packages/accounting/src/main.ts
@@ -3,12 +3,14 @@ import bodyParser from 'body-parser';
 import { run as runReactions } from './reactions';
 import { getTotalStonksForDate } from './db';
 import { consumer, producer } from './kafka';
+import { makeAuthMiddleware, useCanRole } from '../../utils/src/lib/auth';
+import { ROLE_ACCOUNTANT, ROLE_ADMIN } from '@monorepo/kafka-users-common';
 
 const app = express();
 
 app.use(bodyParser.json());
 
-const port = process.env.PORT || 3333;
+const port = parseInt(process.env.PORT || '3333', 10);
 const server = app.listen(port, async () => {
   await runReactions();
   await consumer.connect();
@@ -17,7 +19,9 @@ const server = app.listen(port, async () => {
 });
 server.on('error', console.error);
 
-app.get('/stonks', async (req, res) => {
+const fiefAuthMiddleware = makeAuthMiddleware(app, port);
+
+app.get('/stonks', fiefAuthMiddleware(), useCanRole([ROLE_ACCOUNTANT, ROLE_ADMIN]), async (req, res) => {
   const stonks = getTotalStonksForDate(new Date());
   res.send({ stonks });
 });
