+++ packages/analytics-dashboard/src/main.ts
@@ -2,12 +2,25 @@ import express from 'express';
 import bodyParser from 'body-parser';
 import { run as runReactions } from './reactions';
 import { getMaxPriceForInterval, getTopRevenueToday, loserCountToday } from './ticker';
+import { makeAuthMiddleware, useCanRole } from '../../utils/src/lib/auth';
+import { ROLE_ADMIN } from '@monorepo/kafka-users-common';
 
 const app = express();
 
 app.use(bodyParser.json());
 
-app.get('/winnerPrice/:dayFrom/:dayTo', (req, res) => {
+const port = parseInt(process.env.PORT || '3333', 10);
+const server = app.listen(port, async () => {
+  await runReactions();
+  console.log(`Listening at http://localhost:${port}/api`);
+});
+server.on('error', console.error);
+
+const fiefAuthMiddleware = makeAuthMiddleware(app, port);
+
+
+
+app.get('/winnerPrice/:dayFrom/:dayTo', fiefAuthMiddleware(), useCanRole([ROLE_ADMIN]), (req, res) => {
   const dayFrom = new Date(req.params.dayFrom);
   const dayTo = new Date(req.params.dayTo);
   res.json({
@@ -15,21 +28,15 @@ app.get('/winnerPrice/:dayFrom/:dayTo', (req, res) => {
   });
 });
 
-app.get('/loserCount', (req, res) => {
+app.get('/loserCount', fiefAuthMiddleware(), useCanRole([ROLE_ADMIN]), (req, res) => {
   res.json({
     loserCount: loserCountToday(),
   });
 });
 
-app.get('/topRevenue', (req, res) => {
+app.get('/topRevenue', fiefAuthMiddleware(), useCanRole([ROLE_ADMIN]), (req, res) => {
   res.json({
     topRevenue: getTopRevenueToday(),
   });
 });
 
-const port = process.env.PORT || 3333;
-const server = app.listen(port, async () => {
-  await runReactions();
-  console.log(`Listening at http://localhost:${port}/api`);
-});
-server.on('error', console.error);
